requirements.txt
asgiref                       3.4.1
certifi                       2021.10.8
cffi                          1.15.0
charset-normalizer            2.0.9
coreapi                       2.3.3
coreschema                    0.0.4
cryptography                  36.0.0
defusedxml                    0.7.1
django                        3.2.9
django-allauth                0.40.0
django-filter                 21.1
django-rest-auth              0.9.5
django-rest-swagger           2.2.0
djangorestframework           3.12.4
djangorestframework-simplejwt 5.0.0
idna                          3.3
importlib-metadata            4.8.2
itypes                        1.2.0
Jinja2                        3.0.3
Markdown                      3.3.6
MarkupSafe                    2.0.1
mysqlclient                   2.1.0
oauthlib                      3.1.1
openapi-codec                 1.3.2
pip                           21.1.3
pycparser                     2.21
PyJWT                         2.3.0
python3-openid                3.2.0
pytz                          2021.3
PyYAML                        5.1.2
requests                      2.26.0
requests-oauthlib             1.3.0
setuptools                    56.0.0
simplejson                    3.17.6
six                           1.16.0
sqlparse                      0.4.2
uritemplate                   4.1.1
urllib3                       1.26.7
zipp                          3.6.0